# openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: PT Champion API
  description: API for the PT Champion fitness tracking application
servers:
  - url: /api/v1
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
  schemas:
...
    LocalLeaderboardResponse:
      type: array
      items:
        type: object
        properties:
          userId:
            type: integer
          username:
            type: string
          displayName:
            type: string
            nullable: true
          exerciseId:
            type: integer
          score:
            type: integer
          distanceMeters:
            type: number
          lastUpdated:
            type: string
        required:
          - userId
          - username
          - displayName
          - exerciseId
          - score
...
paths:
...
  /leaderboards/local:
    get:
      summary: Get leaderboard filtered by proximity to user location
      tags:
        - Leaderboard
      operationId: getLocalLeaderboard
      security:
        - BearerAuth: []
      parameters:
        - name: exercise_id
          in: query
          schema:
            type: integer
          description: ID of the exercise to filter leaderboard by
          required: true
        - name: latitude
          in: query
          schema:
            type: number
            format: double
          description: User's current latitude
          required: true
        - name: longitude
          in: query
          schema:
            type: number
            format: double
          description: User's current longitude
          required: true
        - name: radius_meters
          in: query
          schema:
            type: number
            format: double
            default: 8047
          description: Search radius in meters
          required: false
      responses:
        '200':
          description: Local leaderboard retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocalLeaderboardResponse'
        '400':
          description: Missing or invalid required query parameters
        '500':
          description: Internal Server Error retrieving local leaderboard
...